requirements.txt
aiogram==2.21
celery==4.1
python-dotenv
redis
